refactor: 优化项目目录结构
