+++ src/alist_mikananirss/core/__init__.py
@@ -1,6 +1,7 @@
-from .download_manager import DownloadManager
-from .filters import RegexFilter
+from .bot_assistant import BotAssistant
+from .download_manager import *
+from .filter import RegexFilter
 from .notification_sender import NotificationSender
-from .remapper import Remapper, RemapperManager
+from .remapper import RemapFrom, Remapper, RemapperManager, RemapTo
 from .renamer import AnimeRenamer
 from .rss_monitor import RssMonitor
