+++ src/alist_mikananirss/core/renamer.py
@@ -4,7 +4,7 @@ import os
 from loguru import logger
 
 from alist_mikananirss.alist import Alist
-from alist_mikananirss.websites import ResourceInfo
+from alist_mikananirss.websites.models import ResourceInfo
 
 from ..utils import Singleton
