+++ src/alist_mikananirss/websites/__init__.py
@@ -2,5 +2,4 @@ from .acgrip import AcgRip
 from .base import Website, WebsiteFactory
 from .default import DefaultWebsite
 from .dmhy import Dmhy
-from .entities import FeedEntry, ResourceInfo
 from .mikan import Mikan
