+++ src/alist_mikananirss/websites/acgrip.py
@@ -5,9 +5,9 @@ import aiohttp
 import bs4
 
 from alist_mikananirss.extractor import Extractor
+from alist_mikananirss.websites.models import FeedEntry, ResourceInfo
 
 from .base import Website
-from .entities import FeedEntry, ResourceInfo
 
 
 class AcgRip(Website):
