+++ src/alist_mikananirss/websites/dmhy.py
@@ -4,9 +4,9 @@ import aiohttp
 import bs4
 
 from alist_mikananirss.extractor import Extractor
+from alist_mikananirss.websites.models import FeedEntry, ResourceInfo
 
 from .base import Website
-from .entities import FeedEntry, ResourceInfo
 
 
 class Dmhy(Website):
