+++ tests/common/test_database.py
@@ -5,7 +5,7 @@ import pytest
 import pytest_asyncio
 
 from alist_mikananirss.common.database import SubscribeDatabase, db_dirpath
-from alist_mikananirss.websites import ResourceInfo
+from alist_mikananirss.websites.models import ResourceInfo
 
 
 @pytest_asyncio.fixture
