+++ tests/websites/test_dmhy.py
@@ -4,8 +4,8 @@ import feedparser
 import pytest
 
 from alist_mikananirss.extractor import ResourceTitleExtractResult, VideoQuality
-from alist_mikananirss.websites import FeedEntry, ResourceInfo
 from alist_mikananirss.websites.dmhy import Dmhy
+from alist_mikananirss.websites.models import FeedEntry, ResourceInfo
 
 
 @pytest.fixture
